# Fix metric SDK race condition