Discovered while testing and documenting #265, this race condition between `LoadOrStore` and the assignment `rec.recorder = i.meter.exporter.AggregatorFor(rec)` could have resulted in lost updates. Fixing this would require two allocations prior to the `LoadOrStore`. Instead, a first call to `Load` checks for an existing record, followed by the `LoadOrStore` case. This results in fewer allocations overall, as shown in the updated benchmark. 

Before:
```
BenchmarkAcquireNewHandle-8               	 1000000	      1192 ns/op	     275 B/op	       5 allocs/op
BenchmarkAcquireExistingHandle-8          	 2201809	       612 ns/op	     112 B/op	       2 allocs/op
BenchmarkAcquireReleaseExistingHandle-8   	 2219748	       590 ns/op	     112 B/op	       2 allocs/op
```

After:
```
BenchmarkAcquireNewHandle-8               	 1194990	      1157 ns/op	     255 B/op	       5 allocs/op
BenchmarkAcquireExistingHandle-8          	 5107620	       248 ns/op	       0 B/op	       0 allocs/op
BenchmarkAcquireReleaseExistingHandle-8   	 5356180	       253 ns/op	       0 B/op	       0 allocs/op
```